fix: make a few changes to comment and order APIs and related data

+++ apps/shop-corner-api/src/libs/dto/comment/comment.ts
@@ -1,4 +1,4 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, ObjectType } from '@nestjs/graphql';
 import type { ObjectId } from 'mongoose';
 import { CommentGroup, CommentStatus } from '../../enums/comment.enum';
 import { Member, TotalCounter } from '../member/member';
@@ -21,7 +21,7 @@ export class Comment {
 	@Field(() => String)
 	commentRefId: ObjectId;
 
-	@Field(() => Int, { nullable: true })
+	@Field(() => Float, { nullable: true })
 	commentRating?: number;
 
 	@Field(() => String)
